+++ lib/active_record/connection_adapters/sqlserver_adapter.rb
@@ -171,6 +171,10 @@ module ActiveRecord
         true
       end
 
+      def supports_check_constraints?
+        true
+      end
+
       def supports_json?
         version_year >= 2016
       end